refactor: move latest-fork functionality to its own pytest plugin
