+++ conftest.py
@@ -12,11 +12,7 @@ from typing import Any, Dict, List, Tuple, Type
 
 import pytest
 
-from ethereum_test_forks import (
-    ArrowGlacier,
-    InvalidForkError,
-    set_latest_fork_by_name,
-)
+from ethereum_test_forks import ArrowGlacier
 from ethereum_test_tools import (
     BaseTest,
     BlockchainTest,
@@ -72,55 +68,6 @@ def pytest_addoption(parser):
         default="./out/",
         help="Directory to store filled test fixtures",
     )
-    group.addoption(
-        "--latest-fork",
-        action="store",
-        dest="latest_fork",
-        default=None,
-        help="Latest fork used to fill tests",
-    )
-
-
-def pytest_configure(config):
-    """
-    Check parameters and make session-wide configuration changes, such as
-    setting the latest fork.
-    """
-    latest_fork = config.getoption("latest_fork")
-    if latest_fork is not None:
-        try:
-            set_latest_fork_by_name(latest_fork)
-        except InvalidForkError as e:
-            pytest.exit(f"Error applying --latest-fork={latest_fork}: {e}.")
-        except Exception as e:
-            raise e
-    return None
-
-
-@pytest.hookimpl(trylast=True)
-def pytest_report_header(config, start_path):
-    """A pytest hook called to obtain the report header."""
-    bold = "\033[1m"
-    warning = "\033[93m"
-    reset = "\033[39;49m"
-    if config.getoption("latest_fork") is None:
-        header = [
-            (
-                bold
-                + warning
-                + "Only executing fillers with stable/deployed forks: "
-                "Specify an upcoming fork via --latest-fork=fork to "
-                "run experimental fillers." + reset
-            )
-        ]
-    else:
-        header = [
-            (
-                bold + "Executing fillers up to and including "
-                f"{config.getoption('latest_fork')}." + reset
-            ),
-        ]
-    return header
 
 
 @pytest.fixture(autouse=True, scope="session")
